WIP/dsoo/readingfileMD
