+++ helper.js
@@ -1,5 +1,5 @@
 import process from 'node:process';
 
 export const CLI = () => {
-    console.log(process.argv)
+    console.log(process.argv[2])
 }
